wimbee

+++ nextjs-app/components/Navigation.jsx
@@ -20,13 +20,13 @@ function Navigation({ menu, content }) {
   useEffect(() => {
     const header = headerRef.current;
     const homeHero = document.querySelector(".hero-section");
+
     let stInstance;
 
-    if (homeHero) {
-      // If there's a hero section, start with transparent background
-      header.classList.add("bg-transparent");
-      header.classList.remove("bg-light-300");
+    header.classList.add("bg-transparent");
+    header.classList.remove("bg-light-300");
 
+    if (homeHero) {
       stInstance = ScrollTrigger.create({
         trigger: homeHero,
         start: "bottom top",
@@ -44,54 +44,52 @@ function Navigation({ menu, content }) {
           }
         },
       });
+
+      // Fix Safari issues
+      setTimeout(() => {
+        ScrollTrigger.refresh();
+      }, 500);
     } else {
-      // If there's no hero section, use light-300 background
+      // Apply bg-light-300 on all other pages
       header.classList.remove("bg-transparent");
       header.classList.add("bg-light-300");
     }
 
     return () => {
-      // Cleanup the ScrollTrigger instance
-      if (stInstance) {
-        stInstance.kill();
-      }
+      if (stInstance) stInstance.kill();
     };
   }, []);
 
   // Effect for hiding/showing the header on scroll
-  useEffect(() => {
-    let lastScrollTop = 0;
-    let headerHeight = 0;
+  // useEffect(() => {
+  //   let lastScrollTop = 0;
 
-    const handleScroll = () => {
-      const scrollTop =
-        window.pageYOffset || document.documentElement.scrollTop;
+  //   const handleScroll = () => {
+  //     const scrollTop = document.documentElement.scrollTop;
 
-      // Check if the screen width is greater than 768px
-      if (window.innerWidth > 768) {
-        if (scrollTop > lastScrollTop) {
-          // Scrolling down
-          if (headerRef.current) {
-            headerHeight = headerRef.current.getBoundingClientRect().height;
-            headerRef.current.style.top = `-${headerHeight}px`;
-          }
-        } else {
-          // Scrolling up
-          if (headerRef.current) {
-            headerRef.current.style.top = `0px`;
-          }
-        }
-      }
+  //     if (window.innerWidth > 768) {
+  //       if (scrollTop > lastScrollTop) {
+  //         // Hide header
+  //         if (headerRef.current) {
+  //           headerRef.current.style.transform = "translateY(-100%)";
+  //         }
+  //       } else {
+  //         // Show header
+  //         if (headerRef.current) {
+  //           headerRef.current.style.transform = "translateY(0)";
+  //         }
+  //       }
+  //     }
 
-      lastScrollTop = scrollTop;
-    };
+  //     lastScrollTop = scrollTop;
+  //   };
 
-    window.addEventListener("scroll", handleScroll);
+  //   window.addEventListener("scroll", handleScroll, { passive: true });
 
-    return () => {
-      window.removeEventListener("scroll", handleScroll);
-    };
-  }, []);
+  //   return () => {
+  //     window.removeEventListener("scroll", handleScroll);
+  //   };
+  // }, []);
 
   return (
     <header
